agregue dtos de tarea y registro, con sus servicios y modifique controladores para que los incluyan

+++ src/main/java/com/metalurgica1/metalurgica1/repositorio/IEmpleadoRepository.java
@@ -3,5 +3,5 @@ package com.metalurgica1.metalurgica1.repositorio;
 import com.metalurgica1.metalurgica1.modelo.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IEmpleadoRepository extends JpaRepository<Empleado, Integer> {
+public interface IEmpleadoRepository extends JpaRepository<Empleado, Long> {
 }
